работает показ формы ввода коммента после авторизации
