+++ index.html
@@ -14,31 +14,32 @@
       </ul>
       <div class="customer-form" id="customer-form">
         Чтобы добавить комментарий, <u><span class="show-start-form" id="show-start-form">авторизуйтесь</span></u>
+        <div class="add-form invisible" id="add-form" >     
+          <!-- <input
+             type="text"
+             class="add-form-name"
+             placeholder="Введите ваше имя"
+             readonly
+         />
+         <textarea
+             type="textarea"
+             class="add-form-text"
+             placeholder="Введите ваш коментарий"
+             rows="4"
+         ></textarea>
+         <div class="add-form-row">
+             <button class="add-form-button" >Написать</button>
+             </div>
+             </div> 
+         <div class="loaderPost"> 
+             <br> Коментарий добавляется...
+         </div>   
+        
+         -->
+      </div>
       </div>
       
-      <div class="add-form" id="add-form">     
-        <!-- <input
-           type="text"
-           class="add-form-name"
-           placeholder="Введите ваше имя"
-           readonly
-       />
-       <textarea
-           type="textarea"
-           class="add-form-text"
-           placeholder="Введите ваш коментарий"
-           rows="4"
-       ></textarea>
-       <div class="add-form-row">
-           <button class="add-form-button" >Написать</button>
-           </div>
-           </div> 
-       <div class="loaderPost"> 
-           <br> Коментарий добавляется...
-       </div>   
       
-       -->
-    </div>
   </body>
 
 <!-- Начала писать код -->
